Set up player controller and main camera
Car is super tweaked up yo. Diggin the drifting settings and adjusted the speed to feel like 50cc mario kart. (lot of noobs in this class are going to have to play this game and I don't want the only feedback I get to be "I found it hard to stay on the track").